requirements.txt
# Core dependencies
streamlit>=1.28.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0

# Vision-Language Model
open-clip-torch>=2.20.0

# Machine Learning utilities
scikit-learn>=1.3.0

# Visualization
plotly>=5.17.0

# Optional: for better performance
# torch-audio  # Uncomment if needed

